request_threadsafe -> request
